--- booquest-api/src/main/java/com/booquest/booquest_api/auth/adapter/in/web/dto/SocialLoginRequest.java
@@ -1,3 +1,0 @@
-package com.booquest.booquest_api.auth.adapter.in.web.dto;
-
-public record SocialLoginRequest(String accessToken) {}
