--- booquest-api/src/main/java/com/booquest/booquest_api/auth/application/port/in/SocialLoginUseCase.java
@@ -1,7 +1,0 @@
-package com.booquest.booquest_api.auth.application.port.in;
-
-import com.booquest.booquest_api.auth.adapter.in.web.dto.SocialLoginResponse;
-
-public interface SocialLoginUseCase {
-    SocialLoginResponse login(String accessToken);
-}
