+++ booquest-api/src/main/resources/db/migration/V1__init.sql
@@ -21,7 +21,8 @@ CREATE TABLE users (
 
 -- onboarding_profiles
 CREATE TABLE onboarding_profiles (
-                                     user_id    BIGINT PRIMARY KEY REFERENCES users(id) ON DELETE RESTRICT,
+                                     id         BIGSERIAL PRIMARY KEY,
+                                     user_id    BIGINT NOT NULL REFERENCES users(id) ON DELETE RESTRICT,
                                      metadata   JSONB,
                                      created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
                                      updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
